backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "bs4>=0.0.2",
    "fastapi>=0.121.3",
    "langchain>=1.0.8",
    "langchain-anthropic>=1.1.0",
    "langgraph>=1.0.3",
    "pydantic-settings>=2.12.0",
    "lxml>=6.0.2",
    "mistralai>=1.2.6",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
